interface refactor
