+++ frontend/assets/interfaces/index.tsx
@@ -1,0 +1,4 @@
+export * from '@/assets/interfaces/Component';
+export * from '@/assets/interfaces/Data';
+export * from '@/assets/interfaces/Event';
+export * from '@/assets/interfaces/Filter';
