+++ frontend/components/UI/Toggle.tsx
@@ -1,5 +1,5 @@
 import Switch from '@mui/material/Switch';
-import { ToggleType } from '@/assets/interfaces/Component';
+import { ToggleType } from '@/assets/interfaces';
 
 export default function ToggleComponent ({ switchLabel, hide, setHide }: ToggleType) {
   const aria = { inputProps: { 'aria-label': `${switchLabel} Switch` } };
